Let the download url be opened in a new window
When we download a file from, for example Word, it redirects to a new tab. When a user has "Open Websites in the Same Window" option selected, they will see a blank page in the window after they download something. This commit lets these types of download urls to be opened in new tabs, even if the user has selected to open all the urls in same window.